+++ CMakeLists.txt
@@ -2435,6 +2435,7 @@ foreach(_hdr
   include/grpc/channel_credentials/google_default.h
   include/grpc/channel_credentials/local.h
   include/grpc/channel_credentials/ssl.h
+  include/grpc/channel_credentials/tls.h
   include/grpc/compression.h
   include/grpc/event_engine/endpoint_config.h
   include/grpc/event_engine/event_engine.h
@@ -3045,6 +3046,7 @@ foreach(_hdr
   include/grpc/channel_credentials/google_default.h
   include/grpc/channel_credentials/local.h
   include/grpc/channel_credentials/ssl.h
+  include/grpc/channel_credentials/tls.h
   include/grpc/compression.h
   include/grpc/event_engine/endpoint_config.h
   include/grpc/event_engine/event_engine.h
@@ -4527,6 +4529,7 @@ foreach(_hdr
   include/grpc/channel_credentials/google_default.h
   include/grpc/channel_credentials/local.h
   include/grpc/channel_credentials/ssl.h
+  include/grpc/channel_credentials/tls.h
   include/grpc/compression.h
   include/grpc/event_engine/endpoint_config.h
   include/grpc/event_engine/event_engine.h
