+++ Makefile
@@ -1671,6 +1671,7 @@ PUBLIC_HEADERS_C += \
     include/grpc/channel_credentials/google_default.h \
     include/grpc/channel_credentials/local.h \
     include/grpc/channel_credentials/ssl.h \
+    include/grpc/channel_credentials/tls.h \
     include/grpc/compression.h \
     include/grpc/event_engine/endpoint_config.h \
     include/grpc/event_engine/event_engine.h \
@@ -2141,6 +2142,7 @@ PUBLIC_HEADERS_C += \
     include/grpc/channel_credentials/google_default.h \
     include/grpc/channel_credentials/local.h \
     include/grpc/channel_credentials/ssl.h \
+    include/grpc/channel_credentials/tls.h \
     include/grpc/compression.h \
     include/grpc/event_engine/endpoint_config.h \
     include/grpc/event_engine/event_engine.h \
