+++ build_autogenerated.yaml
@@ -261,6 +261,7 @@ libs:
   - include/grpc/channel_credentials/google_default.h
   - include/grpc/channel_credentials/local.h
   - include/grpc/channel_credentials/ssl.h
+  - include/grpc/channel_credentials/tls.h
   - include/grpc/compression.h
   - include/grpc/event_engine/endpoint_config.h
   - include/grpc/event_engine/event_engine.h
@@ -1842,6 +1843,7 @@ libs:
   - include/grpc/channel_credentials/google_default.h
   - include/grpc/channel_credentials/local.h
   - include/grpc/channel_credentials/ssl.h
+  - include/grpc/channel_credentials/tls.h
   - include/grpc/compression.h
   - include/grpc/event_engine/endpoint_config.h
   - include/grpc/event_engine/event_engine.h
@@ -3381,6 +3383,7 @@ libs:
   - include/grpc/channel_credentials/google_default.h
   - include/grpc/channel_credentials/local.h
   - include/grpc/channel_credentials/ssl.h
+  - include/grpc/channel_credentials/tls.h
   - include/grpc/compression.h
   - include/grpc/event_engine/endpoint_config.h
   - include/grpc/event_engine/event_engine.h
