+++ include/grpc/module.modulemap
@@ -9,6 +9,7 @@ header "byte_buffer.h"
   header "channel_credentials/google_default.h"
   header "channel_credentials/local.h"
   header "channel_credentials/ssl.h"
+  header "channel_credentials/tls.h"
   header "compression.h"
   header "fork.h"
   header "grpc.h"
