+++ include/grpcpp/security/tls_certificate_provider.h
@@ -20,6 +20,7 @@
 #include <memory>
 #include <vector>
 
+#include <grpc/channel_credentials/tls.h>
 #include <grpc/grpc_security.h>
 #include <grpc/grpc_security_constants.h>
 #include <grpc/status.h>
