+++ include/grpcpp/security/tls_credentials_options.h
@@ -22,6 +22,7 @@
 #include <memory>
 #include <vector>
 
+#include <grpc/channel_credentials/tls.h>
 #include <grpc/grpc_security.h>
 #include <grpc/grpc_security_constants.h>
 #include <grpc/status.h>
