+++ package.xml
@@ -36,6 +36,7 @@
     <file baseinstalldir="/" name="include/grpc/channel_credentials/google_default.h" role="src" />
     <file baseinstalldir="/" name="include/grpc/channel_credentials/local.h" role="src" />
     <file baseinstalldir="/" name="include/grpc/channel_credentials/ssl.h" role="src" />
+    <file baseinstalldir="/" name="include/grpc/channel_credentials/tls.h" role="src" />
     <file baseinstalldir="/" name="include/grpc/compression.h" role="src" />
     <file baseinstalldir="/" name="include/grpc/event_engine/endpoint_config.h" role="src" />
     <file baseinstalldir="/" name="include/grpc/event_engine/event_engine.h" role="src" />
