+++ src/core/ext/xds/google_mesh_ca_certificate_provider_factory.h
@@ -26,6 +26,7 @@
 #include <string>
 #include <vector>
 
+#include <grpc/channel_credentials/tls.h>
 #include <grpc/grpc_security.h>
 
 #include "src/core/lib/gprpp/ref_counted_ptr.h"
