+++ src/core/lib/security/credentials/ssl/ssl_credentials.cc
@@ -29,6 +29,7 @@
 #include "absl/types/optional.h"
 
 #include <grpc/channel_credentials/ssl.h>
+#include <grpc/channel_credentials/tls.h>
 #include <grpc/impl/codegen/grpc_types.h>
 #include <grpc/support/alloc.h>
 #include <grpc/support/log.h>
