+++ src/core/lib/security/credentials/tls/grpc_tls_certificate_provider.h
@@ -29,6 +29,7 @@
 #include "absl/strings/string_view.h"
 #include "absl/types/optional.h"
 
+#include <grpc/channel_credentials/tls.h>
 #include <grpc/grpc_security.h>
 #include <grpc/support/log.h>
 #include <grpc/support/sync.h>
