+++ src/core/lib/security/credentials/tls/grpc_tls_certificate_verifier.cc
@@ -25,6 +25,7 @@
 
 #include "absl/strings/string_view.h"
 
+#include <grpc/channel_credentials/tls.h>
 #include <grpc/support/alloc.h>
 #include <grpc/support/log.h>
 #include <grpc/support/string_util.h>
