+++ src/core/lib/security/credentials/tls/tls_credentials.h
@@ -21,6 +21,7 @@
 
 #include <grpc/support/port_platform.h>
 
+#include <grpc/channel_credentials/tls.h>
 #include <grpc/grpc.h>
 #include <grpc/grpc_security.h>
