+++ src/core/lib/security/credentials/xds/xds_credentials.cc
@@ -23,6 +23,7 @@
 #include "absl/strings/string_view.h"
 #include "absl/types/optional.h"
 
+#include <grpc/channel_credentials/tls.h>
 #include <grpc/grpc_security_constants.h>
 #include <grpc/impl/codegen/grpc_types.h>
 #include <grpc/support/log.h>
