+++ src/core/lib/security/credentials/xds/xds_credentials.h
@@ -30,6 +30,7 @@
 
 #include "absl/status/status.h"
 
+#include <grpc/channel_credentials/tls.h>
 #include <grpc/grpc.h>
 #include <grpc/grpc_security.h>
