+++ src/core/lib/security/security_connector/tls/tls_security_connector.h
@@ -29,6 +29,7 @@
 #include "absl/strings/string_view.h"
 #include "absl/types/optional.h"
 
+#include <grpc/channel_credentials/tls.h>
 #include <grpc/grpc.h>
 #include <grpc/grpc_security.h>
