+++ src/cpp/server/secure_server_credentials.cc
@@ -27,6 +27,7 @@
 #include <grpc/channel_credentials/alts.h>
 #include <grpc/channel_credentials/local.h>
 #include <grpc/channel_credentials/ssl.h>
+#include <grpc/channel_credentials/tls.h>
 #include <grpc/grpc_security_constants.h>
 #include <grpc/slice.h>
 #include <grpc/status.h>
