+++ test/core/end2end/fixtures/h2_tls_common.h
@@ -26,6 +26,7 @@
 
 #include "absl/strings/string_view.h"
 
+#include <grpc/channel_credentials/tls.h>
 #include <grpc/grpc.h>
 #include <grpc/grpc_security.h>
 #include <grpc/grpc_security_constants.h>
