+++ tools/doxygen/Doxyfile.core.internal
@@ -814,6 +814,7 @@ include/grpc/channel_credentials/alts.h \
 include/grpc/channel_credentials/google_default.h \
 include/grpc/channel_credentials/local.h \
 include/grpc/channel_credentials/ssl.h \
+include/grpc/channel_credentials/tls.h \
 include/grpc/compression.h \
 include/grpc/event_engine/endpoint_config.h \
 include/grpc/event_engine/event_engine.h \
